requirements.txt
pandas
tabula-py=2.10.0
jpype1=1.6.0